requirements.txt
Django==1.6.2
Jinja2==2.7.2
Markdown==2.4
MarkupSafe==0.19
Pillow==2.3.0
South==0.8.4
Werkzeug==0.9.4
beautifulsoup4==4.3.2
bleach==1.4
django-braces==1.4.0
django-common-helpers==0.6.0
django-cron==0.3.3
django-debug-toolbar==1.0.1
django-extensions==1.3.3
django-pipeline==1.3.23
django-sky-thumbnails==0.2.5
django-widget-tweaks==1.3
futures==2.1.6
html5lib==0.999
ipython==1.2.1
jingo==0.7
jsonfield==0.9.20
pytz==2014.1
readline==6.2.4.1
requests==2.2.1
six==1.6.1
socialoauth==0.3.2
sqlparse==0.1.11
wsgiref==0.1.2
